+++ src/main/java/com/liquibaseproject/api/OrdersPathApiController.java
@@ -1,0 +1,47 @@
+package com.liquibaseproject.api;
+
+import com.liquibaseproject.model.ApiResponseSchema;
+import com.liquibaseproject.model.NewOrder;
+import com.liquibaseproject.model.Orders;
+import java.util.UUID;
+
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.CookieValue;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RequestPart;
+import org.springframework.web.multipart.MultipartFile;
+
+import jakarta.validation.constraints.*;
+import jakarta.validation.Valid;
+
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import jakarta.annotation.Generated;
+
+@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.7.0")
+@Controller
+@RequestMapping("${openapi.liquibaseProjectOpenAPI30.base-path:}")
+public class OrdersPathApiController implements OrdersPathApi {
+
+    private final OrdersPathApiDelegate delegate;
+
+    public OrdersPathApiController(@Autowired(required = false) OrdersPathApiDelegate delegate) {
+        this.delegate = Optional.ofNullable(delegate).orElse(new OrdersPathApiDelegate() {});
+    }
+
+    @Override
+    public OrdersPathApiDelegate getDelegate() {
+        return delegate;
+    }
+
+}